+++ RoomMates/ContentView.swift
@@ -85,24 +85,6 @@ struct ContentView: View {
     }
 }
 
-
-struct ProfileView: View {
-    @EnvironmentObject var authViewModel: AuthViewModel
-    
-    var body: some View {
-        Button(action: {
-            authViewModel.signOut()
-        }) {
-            Text("Log Out")
-                .frame(maxWidth: 300)
-                .padding()
-                .background(Color.red)
-                .foregroundColor(.white)
-                .clipShape(RoundedRectangle(cornerRadius: 10))
-        }
-    }
-}
-
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
         ContentView()
